Split out sample selection strategies into actions
