+++ src/reducers/profile-view.js
@@ -184,34 +184,44 @@ function viewOptionsPerThread(
       });
     }
     case 'CHANGE_SELECTED_CALL_NODE': {
-      const { selectedCallNodePath, threadIndex } = action;
+      const {
+        selectedCallNodePath,
+        threadIndex,
+        optionalExpandedToCallNodePath,
+      } = action;
 
       const threadState = state[threadIndex];
       const previousSelectedCallNodePath = threadState.selectedCallNodePath;
 
       // If the selected node doesn't actually change, let's return the previous
       // state to avoid rerenders.
-      if (arePathsEqual(selectedCallNodePath, previousSelectedCallNodePath)) {
+      if (
+        arePathsEqual(selectedCallNodePath, previousSelectedCallNodePath) &&
+        !optionalExpandedToCallNodePath
+      ) {
         return state;
       }
 
       let { expandedCallNodePaths } = threadState;
+      const expandToNode = optionalExpandedToCallNodePath
+        ? optionalExpandedToCallNodePath
+        : selectedCallNodePath;
 
       /* Looking into the current state to know whether we want to generate a
        * new one. It can be expensive to clone when we have a lot of expanded
        * lines, but it's very infrequent that we actually want to expand new
        * lines as a result of a selection. */
-      const selectedNodeParentPaths = [];
-      for (let i = 1; i < selectedCallNodePath.length; i++) {
-        selectedNodeParentPaths.push(selectedCallNodePath.slice(0, i));
+      const expandToNodeParentPaths = [];
+      for (let i = 1; i < expandToNode.length; i++) {
+        expandToNodeParentPaths.push(expandToNode.slice(0, i));
       }
-      const hasNewExpandedPaths = selectedNodeParentPaths.some(
+      const hasNewExpandedPaths = expandToNodeParentPaths.some(
         path => !expandedCallNodePaths.has(path)
       );
 
       if (hasNewExpandedPaths) {
         expandedCallNodePaths = new PathSet(expandedCallNodePaths);
-        selectedNodeParentPaths.forEach(path =>
+        expandToNodeParentPaths.forEach(path =>
           expandedCallNodePaths.add(path)
         );
       }
